server/requirements.txt
aiohttp
scipy
tqdm
numpy
tempfile
ffmpeg-python
